requirements.txt
pymupdf==1.24.10
pytesseract==0.3.10
Pillow==9.3.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.83.0
gspread==5.7.2
